Changed streaming data over promises
It would be great to have real time data all the time, but in the long run it will be more effords and xost so I decided to switched to a promised base system. Everytime the user enters the pages loads the content and also the last 4 jobs from Firestore, then you can load more if needed.