Routing improved in the individual project close action
Before when you click the close button in the individual project page it will send you to the '/works' page. But now it will check if you come from an external page or in domain and will redirect you depending of that. If page before it's external it will send you to the '/works' pages as before. However, if the previous page it's one inside the domain it will send you to that previous page.